# src/docs/openapi3_0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger CRUD Aplication - OpenAPI 3.0",
    "description": "Simples API CRUD parte do desafio que efetua atualizações diárias de produtos retirados da API  https://br.openfoodfacts.org/data",
    "contact": {
      "email": "manoelbatista902@gmail.com",
      "name": "Manoel Batista R Filho"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/v1"
    }
  ],
  "components": {
    "schemas": {
      "product": {
        "type": "object",
        "required": [
          "code",
          "status",
          "imported_t",
          "url",
          "creator",
          "created_t",
          "last_modified_t",
          "product_name",
          "quantity",
          "brands",
          "categories",
          "labels",
          "cities",
          "purchase_places",
          "stores",
          "ingredients_text",
          "traces",
          "serving_size",
          "serving_quantity",
          "nutriscore_score",
          "nutriscore_grade",
          "main_category",
          "image_url"
        ],
        "properties": {
          "code": {
            "type": "number",
            "description": "auto generated field",
            "example": 20221126
          },
          "status": {
            "type": "string",
            "description": "auto generated field",
            "example": "published"
          },
          "imported_t": {
            "type": "string",
            "description": "auto generated field",
            "example": "2020-02-07T16:00:00.000Z"
          },
          "url": {
            "type": "string",
            "description": "auto generated field",
            "example": "https://world.openfoodfacts.org/product/20221126"
          },
          "creator": {
            "type": "string",
            "description": "auto generated field",
            "example": "securita"
          },
          "created_t": {
            "type": "number",
            "description": "auto generated field",
            "example": 1415302075
          },
          "last_modified_t": {
            "type": "number",
            "description": "auto generated field",
            "example": 1572265837
          },
          "product_name": {
            "type": "string",
            "description": "auto generated field",
            "example": "Madalenas quadradas"
          },
          "quantity": {
            "type": "string",
            "description": "auto generated field",
            "example": "380 g (6 x 2 u.)"
          },
          "brands": {
            "type": "string",
            "description": "auto generated field",
            "example": "La Cestera"
          },
          "categories": {
            "type": "string",
            "description": "auto generated field",
            "example": "Lanches comida, Lanches doces, Biscoitos e Bolos, Bolos, Madalenas"
          },
          "labels": {
            "type": "string",
            "description": "auto generated field",
            "example": "Contem gluten, Contém derivados de ovos, Contém ovos"
          },
          "cities": {
            "type": "string",
            "description": "auto generated field",
            "example": null
          },
          "purchase_places": {
            "type": "string",
            "description": "auto generated field",
            "example": "Braga,Portugal"
          },
          "stores": {
            "type": "string",
            "description": "auto generated field",
            "example": "Lidl"
          },
          "ingredients_text": {
            "type": "string",
            "description": "auto generated field",
            "example": "farinha de trigo, açúcar, óleo vegetal de girassol, clara de ovo, ovo, humidificante (sorbitol), levedantes químicos (difosfato dissódico, hidrogenocarbonato de sódio), xarope de glucose-frutose, sal, aroma"
          },
          "traces": {
            "type": "string",
            "description": "auto generated field",
            "example": "Frutos de casca rija,Leite,Soja,Sementes de sésamo,Produtos à base de sementes de sésamo"
          },
          "serving_size": {
            "type": "string",
            "description": "auto generated field",
            "example": "madalena 31.7 g"
          },
          "serving_quantity": {
            "type": "number",
            "description": "auto generated field",
            "example": 31.7
          },
          "nutriscore_score": {
            "type": "number",
            "description": "auto generated field",
            "example": 17
          },
          "nutriscore_grade": {
            "type": "string",
            "description": "auto generated field",
            "example": "d"
          },
          "main_category": {
            "type": "string",
            "description": "auto generated field",
            "example": "en:madeleines"
          },
          "image_url": {
            "type": "string",
            "description": "auto generated field",
            "example": "https://static.openfoodfacts.org/images/products/20221126/front_pt.5.400.jpg"
          }
        }
      },
      "page": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "lista de elementos retornados",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/product"
                }
              ]
            }
          },
          "page": {
            "type": "number",
            "description": "o numeto da pagina atual"
          },
          "total_size": {
            "type": "number",
            "description": "quantidade total de elementos"
          },
          "size": {
            "type": "number",
            "description": "quantidade de elementos desta pagina"
          },
          "sort": {
            "type": "object",
            "properties": {
              "by": {
                "type": "string"
              },
              "sort": {
                "type": "string",
                "enum": [
                  "ASC",
                  "DES"
                ]
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Requisição da pagina de produtos",
        "description": "Requisita uma pagina de tamanho variavel de acordo as query params, note que se houver erro na  operação será enviada uma mensagem padrão do express com o erro e motivo provavel",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "o numero da pagina requisitada",
            "schema": {
              "type": "number",
              "example": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "quantidade maxima de dados a serem requisitados",
            "schema": {
              "type": "number",
              "example": 50
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "ordem da query a ser executada",
            "schema": {
              "type": "number",
              "example": "code,name,ASC"
            }
          }
        ],
        "operationId": "getPage",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/page"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "products"
        ],
        "summary": "adiciona ou modifica o produto",
        "description": "retorna o produto já modificado",
        "operationId": "putProduct",
        "requestBody": {
          "$ref": "#/components/schemas/product"
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/product"
                }
              }
            }
          },
          "400": {
            "description": "Object is null or undefine!"
          }
        }
      }
    },
...
  }
}